Fix non-portable `1ul << 63` in `negate.cpp`
Use `UInt64(1) << 63` instead of `1ul << 63` for portability, since
`unsigned long` may be 32 bits on some platforms (e.g. Windows).
Consistent with the `UInt128(1) << 127` and `UInt256(1) << 255`
patterns on the adjacent lines.

+++ src/Functions/negate.cpp
@@ -131,7 +131,7 @@ template <> struct FunctionUnaryArithmeticMonotonicity<NameNegate>
         {
             switch (f.getType())
             {
-                case Field::Types::UInt64: return f.safeGet<UInt64>() > 1ul << 63;
+                case Field::Types::UInt64: return f.safeGet<UInt64>() > UInt64(1) << 63;
                 case Field::Types::UInt128: return f.safeGet<UInt128>() > UInt128(1) << 127;
                 case Field::Types::UInt256: return f.safeGet<UInt256>() > UInt256(1) << 255;
                 default: break;
